+++ dotnet-scripts.ps1
@@ -1,3 +1,3 @@
 # TODO: Сделать флаг
 
-dotnet run --project .\src\piligrimm-server\Web\Web
+dotnet run --project .\src\Piligrimm.Server\Web\Web
