+++ src/Piligrimm.Server/Models/Models/Interfaces/Category/ICategory.cs
