+++ src/Piligrimm.Server/Web/Web/Market/Category/Category.cs
@@ -1,0 +1,23 @@
+using Microsoft.AspNetCore.Mvc;
+using Service.Market;
+
+namespace Web.Market
+{
+    [Controller]
+    [Route("market/category")]
+    public class ControllerCategory : ControllerBase
+    {
+        private readonly ICategoryService categoryService;
+
+        public ControllerCategory(ICategoryService _categoryService)
+        {
+            categoryService = _categoryService;
+        }
+
+        [HttpGet(Name = "GetAll")]
+        public IEnumerable<Models.Market.Category> GetAll()
+        {
+            return categoryService.GetAll();
+        }
+    }
+}
