--- src/piligrimm-server/Service/Service/Market/Category/ServiceCategory.cs
@@ -1,20 +1,0 @@
-using Application.Market;
-using Models.Market;
-
-namespace Service.Market
-{
-    public class ServiceCategory : IServiceCategory
-    {
-        private readonly IApplicationCategory applicationCategory;
-
-        public ServiceCategory(IApplicationCategory _applicationCategory)
-        {
-            applicationCategory = _applicationCategory;
-        }
-
-        public IEnumerable<Category> GetAll()
-        {
-            return applicationCategory.GetAll();
-        }
-    }
-}
